Extract compilation code from styles-plugin
