+++ packages/wmr/src/plugins/wmr/styles/sass.js
@@ -1,0 +1,9 @@
+import { transformCss } from '../../../lib/transform-css.js';
+
+/**
+ * @param {string} sass
+ * @returns {string} css
+ */
+export function processSass(sass) {
+	return transformCss(sass);
+}
