+++ packages/wmr/src/wmr-middleware.js
@@ -2,7 +2,7 @@ import { resolve, dirname, relative, sep, posix, isAbsolute, normalize, basename
 import { promises as fs, createReadStream } from 'fs';
 import * as kl from 'kolorist';
 import wmrPlugin, { getWmrClient } from './plugins/wmr/plugin.js';
-import wmrStylesPlugin, { modularizeCss, processSass } from './plugins/wmr/styles-plugin.js';
+import wmrStylesPlugin from './plugins/wmr/styles/styles-plugin.js';
 import { createPluginContainer } from './lib/rollup-plugin-container.js';
 import { transformImports } from './lib/transform-imports.js';
 import { normalizeSpecifier } from './plugins/npm-plugin/index.js';
@@ -12,6 +12,8 @@ import { debug, formatPath } from './lib/output-utils.js';
 import { getPlugins } from './lib/plugins.js';
 import { watch } from './lib/fs-watcher.js';
 import { matchAlias, resolveAlias } from './lib/aliasing.js';
+import { modularizeCss } from './plugins/wmr/styles/css-modules.js';
+import { processSass } from './plugins/wmr/styles/sass.js';
 
 const NOOP = () => {};
